+++ caikit/runtime/__init__.py
@@ -13,4 +13,8 @@
 # limitations under the License.
 
 # Local
-from .service_factory import get_inference_request, get_train_params, get_train_request
+from .grpc.service_factory import (
+    get_inference_request,
+    get_train_params,
+    get_train_request,
+)
