+++ caikit/runtime/grpc/grpc_server.py
@@ -30,6 +30,17 @@ import alog
 # Local
 # Get the injectable servicer class definitions
 from caikit import get_config
+from caikit.runtime.grpc.service_factory import ServicePackage, ServicePackageFactory
+from caikit.runtime.grpc.servicers.global_predict_servicer import GlobalPredictServicer
+from caikit.runtime.grpc.servicers.global_train_servicer import GlobalTrainServicer
+from caikit.runtime.grpc.servicers.info_servicer import InfoServicer
+from caikit.runtime.grpc.servicers.model_runtime_servicer import (
+    ModelRuntimeServicerImpl,
+)
+from caikit.runtime.grpc.servicers.model_train_servicer import ModelTrainServicerImpl
+from caikit.runtime.grpc.servicers.training_management_servicer import (
+    TrainingManagementServicerImpl,
+)
 from caikit.runtime.interceptors.caikit_runtime_server_wrapper import (
     CaikitRuntimeServerWrapper,
 )
@@ -39,15 +50,6 @@ from caikit.runtime.protobufs import (
     process_pb2_grpc,
 )
 from caikit.runtime.server_base import RuntimeServerBase
-from caikit.runtime.service_factory import ServicePackage, ServicePackageFactory
-from caikit.runtime.servicers.global_predict_servicer import GlobalPredictServicer
-from caikit.runtime.servicers.global_train_servicer import GlobalTrainServicer
-from caikit.runtime.servicers.info_servicer import InfoServicer
-from caikit.runtime.servicers.model_runtime_servicer import ModelRuntimeServicerImpl
-from caikit.runtime.servicers.model_train_servicer import ModelTrainServicerImpl
-from caikit.runtime.servicers.training_management_servicer import (
-    TrainingManagementServicerImpl,
-)
 
 # Have pylint ignore broad exception catching in this file so that we can log all
 # unexpected errors using alog.
