+++ caikit/runtime/http/http_server.py
@@ -62,16 +62,16 @@ from caikit.config import get_config
 from caikit.core.data_model import DataBase
 from caikit.core.data_model.dataobject import make_dataobject
 from caikit.core.toolkit.sync_to_async import async_wrap_iter
+from caikit.runtime.grpc.service_factory import ServicePackage
+from caikit.runtime.grpc.servicers.global_predict_servicer import GlobalPredictServicer
+from caikit.runtime.grpc.servicers.global_train_servicer import GlobalTrainServicer
+from caikit.runtime.grpc.servicers.info_servicer import InfoServicer
 from caikit.runtime.server_base import RuntimeServerBase
-from caikit.runtime.service_factory import ServicePackage
 from caikit.runtime.service_generation.rpcs import (
     CaikitRPCBase,
     ModuleClassTrainRPC,
     TaskPredictRPC,
 )
-from caikit.runtime.servicers.global_predict_servicer import GlobalPredictServicer
-from caikit.runtime.servicers.global_train_servicer import GlobalTrainServicer
-from caikit.runtime.servicers.info_servicer import InfoServicer
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
 
 ## Globals #####################################################################
