+++ tests/runtime/grpc/servicers/test_global_train_servicer_impl.py
@@ -29,12 +29,12 @@ from caikit.config import get_config
 from caikit.core import MODEL_MANAGER
 from caikit.core.data_model.producer import ProducerId
 from caikit.interfaces.common.data_model.stream_sources import S3Path
-from caikit.runtime.service_factory import (
+from caikit.runtime.grpc.service_factory import (
     get_inference_request,
     get_train_params,
     get_train_request,
 )
-from caikit.runtime.servicers.global_train_servicer import GlobalTrainServicer
+from caikit.runtime.grpc.servicers.global_train_servicer import GlobalTrainServicer
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
 from sample_lib.data_model.sample import (
     OtherOutputType,
