+++ tests/runtime/grpc/servicers/test_servicer_impl_subprocess.py
@@ -25,8 +25,8 @@ import pytest
 # Local
 from .test_model_train_servicer_impl import sample_model_train_servicer
 from caikit.core import MODEL_MANAGER
+from caikit.runtime.grpc.service_factory import get_train_params, get_train_request
 from caikit.runtime.protobufs import process_pb2
-from caikit.runtime.service_factory import get_train_params, get_train_request
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
 from sample_lib.data_model.sample import SampleTrainingType
 from sample_lib.modules import SampleModule
