+++ tests/runtime/grpc/servicers/test_training_management_servicer.py
@@ -35,7 +35,7 @@ from caikit.interfaces.runtime.data_model import (
     TrainingInfoRequest,
     TrainingStatusResponse,
 )
-from caikit.runtime.servicers.training_management_servicer import (
+from caikit.runtime.grpc.servicers.training_management_servicer import (
     TrainingManagementServicerImpl,
 )
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
@@ -72,7 +72,6 @@ class MockModelFuture:
 
 
 def test_training_runs(training_management_servicer, training_pool):
-
     # Create a future and set it in the training manager
     event = threading.Event()
     model_future = MODEL_MANAGER.train(
@@ -232,7 +231,6 @@ def test_training_cancel_on_mock_model_future(training_management_servicer):
 
 
 def test_training_complete_status(training_management_servicer, training_pool):
-
     # Create a future and set it in the training manager
     model_future = MODEL_MANAGER.train(
         SampleModule,
