+++ tests/runtime/grpc/test_grpc_server.py
@@ -52,7 +52,8 @@ from caikit.interfaces.runtime.data_model import (
     TrainingStatusResponse,
 )
 from caikit.runtime import get_inference_request, get_train_params, get_train_request
-from caikit.runtime.grpc_server import RuntimeGRPCServer
+from caikit.runtime.grpc.grpc_server import RuntimeGRPCServer
+from caikit.runtime.grpc.service_factory import ServicePackage, ServicePackageFactory
 from caikit.runtime.model_management.model_manager import ModelManager
 from caikit.runtime.protobufs import (
     model_runtime_pb2,
@@ -60,7 +61,6 @@ from caikit.runtime.protobufs import (
     process_pb2,
     process_pb2_grpc,
 )
-from caikit.runtime.service_factory import ServicePackage, ServicePackageFactory
 from caikit.runtime.utils.servicer_util import build_caikit_library_request_dict
 from sample_lib import CompositeModule, InnerModule, OtherModule, SamplePrimitiveModule
 from sample_lib.data_model import (
@@ -1325,7 +1325,7 @@ def test_grpc_sever_shutdown_with_model_poll(open_port):
     """Test that a SIGINT successfully shuts down the running server"""
     with tempfile.TemporaryDirectory() as workdir:
         server_proc = ModuleSubproc(
-            "caikit.runtime.grpc_server",
+            "caikit.runtime.grpc.grpc_server",
             kill_timeout=30.0,
             RUNTIME_GRPC_PORT=str(open_port),
             RUNTIME_LOCAL_MODELS_DIR=workdir,
