+++ tests/runtime/http/conftest.py
@@ -20,9 +20,9 @@ import alog
 
 # Local
 from caikit.core import MODEL_MANAGER
-from caikit.runtime import http_server
+from caikit.runtime.grpc.servicers.global_predict_servicer import GlobalPredictServicer
+from caikit.runtime.http import http_server
 from caikit.runtime.model_management.loaded_model import LoadedModel
-from caikit.runtime.servicers.global_predict_servicer import GlobalPredictServicer
 from tests.conftest import temp_config
 
 log = alog.use_channel("TEST-HTTPCONF")
