+++ tox.ini
@@ -77,9 +77,9 @@ extras = dev-test
 commands =
     pytest tests/core
 
-# Ensure tests targeting caikit.runtime.grpc_server can be run with no optional dependencies
+# Ensure tests targeting caikit grpc_server can be run with no optional dependencies
 [testenv:grpc_server]
-description = run tests against caikit.runtime.grpc_server without anything else
+description = run tests against caikit grpc_server without anything else
 base = ""
 extras = 
     dev-test
@@ -87,9 +87,9 @@ extras =
 commands =
     pytest tests/runtime/grpc
 
-# Ensure tests targeting caikit.runtime.http_server can be run with no optional dependencies
+# Ensure tests targeting caikit http_server can be run with no optional dependencies
 [testenv:http_server]
-description = run tests against caikit.runtime.http_server without anything else
+description = run tests against caikit http_server without anything else
 base = ""
 extras = 
     dev-test
